Quantum Test and Much More
Removed JMP instruction, it was useless
Added the ability to place labels in MEM objects (this can replace JMP
using "B [:label]")
Working example code added for quantum drive
Re-added the rate counter. It may still be bugged, but now the
calculations are different so we'll see how it goes. Tests glean a rate
of around 1300+ hertz.
B and REG take a MEM rather than a GENERIC